[#175973060] move BpdAmount
