+++ ts/features/bonus/bpd/components/bpdCardComponent/BpdCardComponent.tsx
@@ -5,7 +5,7 @@ import { fromNullable } from "fp-ts/lib/Option";
 import { H2 } from "../../../../../components/core/typography/H2";
 import { H5 } from "../../../../../components/core/typography/H5";
 import { IOColors } from "../../../../../components/core/variables/IOColors";
-import { BpdAmount } from "../../store/actions/amount";
+import { BpdAmount } from "../../saga/networking/amount";
 import { BpdPeriod, BpdPeriodStatus } from "../../store/actions/periods";
 import { H4 } from "../../../../../components/core/typography/H4";
 import I18n from "../../../../../i18n";
